Fade In / Out
I added the fade in / out mechanic that is controlled with the singular button, to do this I used animations because for whatever reason I couldn't change the transparency of the UI image. I also added an event at the end of the fade in animation so I can set up the lose condition later.