Audio, Death, And Random Events
In this commit not only didi I add audio to enhance the players experiance, but I also added a death screen so you actually have a way to die. The random events have been added and happen randomly with the timer I made last time, its only one event so far and the events overlap on top of one another but I hope to tweak that's later.